requirements.txt
streamlit
moviepy
pydub
openai
google-cloud-speech
google-cloud-texttospeech
